requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
gtts==2.5.1
pydantic==2.5.0
transformers==4.36.0
torch==2.1.0
sentencepiece==0.1.99